R6 Class for a Glycoproteomics or Glycomics Experiment

A class managing expression matrix, sample information, and variable information.

Details

This class synchronizes three data types commonly used in a glycproteomics or glycomics research: the expression matrix, the sample information, and the variable information. It helps ensure that the three datasets altered in sync. For example, filtering samples based on some conditions will automatically affect both the expression matrix and the sample information.

Public fields

name

Name of the experiment.

Methods

Public methods

Experiment$new()

Experiment$get_expr_mat()

Experiment$get_sample_info()

Experiment$get_var_info()

Experiment$filter_samples()

Experiment$filter_variables()

Experiment$mutate_samples()

Experiment$mutate_variables()

Experiment$select_samples()

Experiment$select_variables()

Experiment$print()

Experiment$clone()

Method new()

Create a new experiment object.

Usage

Experiment$new(
  name = NA,
  expr_mat = NA,
  sample_info = NA,
  var_info = NA,
  rownames = FALSE
)

Arguments

name

Name of the experiment.

expr_mat

Expression matrix. Rows are variables and columns are samples.

sample_info

Sample information, a tibble or a data.frame. The first column should be "sample", corresponding to the column names of expr_mat. The order of samples doesn't matter.

var_info

Variable information, a tibble or a data.frame. The first column should be "variable", corresponding to the row names of expr_mat. The order of variables doesn't matter.

rownames

Whether rownames of sample_info and var_info should be used. If TRUE, the row names of sample_info and sample_info will be converted to the first column (with names of "sample" and "variable", respectively).

Method get_expr_mat()

Get a copy of the expression matrix. The rows are variables and the columns are samples. A function get_expr_mat() is also provided for this purpose.

Usage

Experiment$get_expr_mat()

Returns

A copy of the expression matrix.

Method get_sample_info()

Get a copy of the sample information tibble. This is a tibble with the first conserved column "sample", and the rest columns as meta data of samples, e.g. "group", "batch", etc. The "sample" column equals to the column names of expr_mat. A function get_sample_info() is also provided for this purpose.

Usage

Experiment$get_sample_info()

Returns

A copy of the sample information tibble.

Method get_var_info()

Get a copy of the variable information tibble. This is a tibble with the first conserved column "variable", and the rest columns as meta data of variables, e.g. "peptide", "protein", "glycan_composition", etc. The "variable" column equals to the row names of expr_mat. A function get_var_info() is also provided for this purpose.

Usage

Experiment$get_var_info()

Returns

A copy of the variable information tibble.

Method filter_samples()

This method filters samples based on conditions specified in the dplyr::filter() function. For example, filter_samples(group == "A") will keep samples with "group" equal to "A".

Usage

Experiment$filter_samples(...)

Arguments

...

Conditions for filtering samples, passed to dplyr::filter().

Details

This will affect both the expression matrix and the sample information tibble. If no samples meet the condition(s), an error will be thrown. The Experiment object will be updated in place. The Experiment object will be returned invisibly to allow chaining.

Returns

The Experiment object.

Method filter_variables()

This method filters variables based on conditions specified in the dplyr::filter() function. For example, filter_variables(type == "B") will keep variables with "type" equal to "B".

Usage

Experiment$filter_variables(...)

Arguments

...

Conditions for filtering variables, passed to dplyr::filter().

Details

This will affect both the expression matrix and the variable information tibble. If no variables meet the condition(s), an error will be thrown. The Experiment object will be updated in place. The Experiment object will be returned invisibly to allow chaining.

Returns

The Experiment object.

Method mutate_samples()

This function mutates the sample information tibble using dplyr::mutate(). For example, mutate_samples(new_group = if_else(new_col = 1) will add a new column "new_group" to the sample information tibble.

Usage

Experiment$mutate_samples(...)

Arguments

...

Mutations for the sample information tibble, passed to dplyr::mutate().

Details

The Experiment object will be updated in place. The Experiment object will be returned invisibly to allow chaining.

The columns "sample" is protected from being modified or renamed. It is essential for keeping the linkage between the sample information and the expression matrix. You may add new columns by transforming the "sample" column.

Returns

The Experiment object.

Method mutate_variables()

This function mutates the variable information tibble using dplyr::mutate(). For example, mutate_variables(new_type = if_else(new_col = 1) will add a new column "new_type" to the variable information tibble.

Usage

Experiment$mutate_variables(...)

Arguments

...

Mutations for the variable information tibble, passed to dplyr::mutate().

Details

The Experiment object will be updated in place. The Experiment object will be returned invisibly to allow chaining.

The columns "variable" is protected from being modified or renamed. It is essential for keeping the linkage between the variable information and the expression matrix. You may add new columns by transforming the "variable" column.

Returns

The Experiment object.

Method select_samples()

This function selects columns in the sample information tibble. You don't need to select the "sample" column explicitly, as it is always selected. For example, exp$select_samples(group) will result in a tibble with the "sample" and "group" columns.

Usage

Experiment$select_samples(...)

Arguments

...

<tidy-select> for selecting columns, passed to dplyr::select().

Details

The Experiment object will be updated in place. The Experiment object will be returned invisibly to allow chaining.

When the user tries to select or deselect the "sample" column explicitly, an error will be thrown.

Returns

The Experiment object.

Method select_variables()

This function selects columns in the variable information tibble. You don't need to select the "variable" column explicitly, as it is always selected. For example, exp$select_variables(peptide, protein) will result in a tibble with the "variable", "peptide", and "protein" columns.

Usage

Experiment$select_variables(...)

Arguments

...

<tidy-select> for selecting columns, passed to dplyr::select().

Details

The Experiment object will be updated in place. The Experiment object will be returned invisibly to allow chaining.

When the user tries to select or deselect the "variable" column explicitly, an error will be thrown.

Returns

The Experiment object.

Method print()

Print the basic information of the Experiment object.

Usage

Experiment$print(...)

Arguments

...

Ignored.

Method clone()

The objects of this class are cloneable with this method.

Usage

Experiment$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.