Filter Samples or Variables of an Experiment

Getting a subset of an experiment() by filtering samples or variables.

The same syntax as dplyr::filter() is used. For example, to get a subset of an experiment keeping only "HC" samples, use filter_samples(exp, group == "HC"). This actually calls dplyr::filter() on the sample information tibble with condition group == "HC", and then updates the expression matrix accordingly.

If no samples or variables are left after filtering, an error is thrown.

Usage

filter_samples(exp, ...)

filter_variables(exp, ...)

Arguments

exp

An experiment().

...

<data-masking> Expression to filter samples or variables. passed to dplyr::filter() internally.

Value

An new experiment() object.

Examples

library(magrittr)

# Create a toy experiment for demonstration
expr_mat <- matrix(1:25, nrow = 5)
colnames(expr_mat) <- paste0("S", 1:5)
rownames(expr_mat) <- paste0("V", 1:5)
sample_info <- tibble::tibble(
  sample = paste0("S", 1:5),
  group = c("A", "A", "A", "B", "B")
)
var_info <- tibble::tibble(
  variable = paste0("V", 1:5),
  type = c("X", "X", "Y", "Y", "Y")
)
exp <- experiment("my_exp", expr_mat, sample_info, var_info)

# Filter samples
sub_exp_1 <- filter_samples(exp, group == "A")
get_sample_info(sub_exp_1)
#> # A tibble: 3 × 2
#>   sample group
#>   <chr>  <chr>
#> 1 S1     A    
#> 2 S2     A    
#> 3 S3     A    
get_expr_mat(sub_exp_1)
#>    S1 S2 S3
#> V1  1  6 11
#> V2  2  7 12
#> V3  3  8 13
#> V4  4  9 14
#> V5  5 10 15

# Filter variables
sub_exp_2 <- filter_variables(exp, type == "X")
get_var_info(sub_exp_2)
#> # A tibble: 2 × 2
#>   variable type 
#>   <chr>    <chr>
#> 1 V1       X    
#> 2 V2       X    
get_expr_mat(sub_exp_2)
#>    S1 S2 S3 S4 S5
#> V1  1  6 11 16 21
#> V2  2  7 12 17 22

# Use pipe
sub_exp_3 <- exp %>%
  filter_samples(group == "A") %>%
  filter_variables(type == "X")
get_sample_info(sub_exp_3)
#> # A tibble: 3 × 2
#>   sample group
#>   <chr>  <chr>
#> 1 S1     A    
#> 2 S2     A    
#> 3 S3     A    
get_var_info(sub_exp_3)
#> # A tibble: 2 × 2
#>   variable type 
#>   <chr>    <chr>
#> 1 V1       X    
#> 2 V2       X    
get_expr_mat(sub_exp_3)
#>    S1 S2 S3
#> V1  1  6 11
#> V2  2  7 12