Mutate Sample or Variable Information

Mutate the sample or variable information tibble of an experiment().

The same syntax as dplyr::mutate() is used. For example, to add a new column to the sample information tibble, use mutate_samples(exp, new_column = value). This actually calls dplyr::mutate() on the sample information tibble with new_column = value.

If the sample column in sample_info or the variable column in var_info is to be modified, the new column must be unique, otherwise an error is thrown. The column names or row names of expr_mat will be updated accordingly.

Usage

mutate_samples(exp, ...)

mutate_variables(exp, ...)

Arguments

exp

An experiment().

...

<data-masking> Name-value pairs, passed to dplyr::mutate() internally.

Value

An new experiment() object.

Examples

# Create a toy experiment for demonstration
expr_mat <- matrix(1:25, nrow = 5)
colnames(expr_mat) <- paste0("S", 1:5)
rownames(expr_mat) <- paste0("V", 1:5)
sample_info <- tibble::tibble(
  sample = paste0("S", 1:5),
  group = c("A", "A", "A", "B", "B")
)
var_info <- tibble::tibble(
  variable = paste0("V", 1:5),
  type = c("X", "X", "Y", "Y", "Y")
)
exp <- experiment("my_exp", expr_mat, sample_info, var_info)

# Add a new column to sample information tibble or variable information tibble
mutate_samples(exp, new_column = c(1, 2, 3, 4, 5))$sample_info
#> # A tibble: 5 × 3
#>   sample group new_column
#>   <chr>  <chr>      <dbl>
#> 1 S1     A              1
#> 2 S2     A              2
#> 3 S3     A              3
#> 4 S4     B              4
#> 5 S5     B              5
mutate_variables(exp, new_column = c("A", "A", "B", "B", "B"))$var_info
#> # A tibble: 5 × 3
#>   variable type  new_column
#>   <chr>    <chr> <chr>     
#> 1 V1       X     A         
#> 2 V2       X     A         
#> 3 V3       Y     B         
#> 4 V4       Y     B         
#> 5 V5       Y     B         

# Modify existing columns
mutate_samples(exp, group = dplyr::if_else(group == "A", "good", "bad"))$sample_info
#> # A tibble: 5 × 2
#>   sample group
#>   <chr>  <chr>
#> 1 S1     good 
#> 2 S2     good 
#> 3 S3     good 
#> 4 S4     bad  
#> 5 S5     bad  
mutate_variables(exp, type = dplyr::if_else(type == "X", "good", "bad"))$var_info
#> # A tibble: 5 × 2
#>   variable type 
#>   <chr>    <chr>
#> 1 V1       good 
#> 2 V2       good 
#> 3 V3       bad  
#> 4 V4       bad  
#> 5 V5       bad  

# Modify the `sample` column in sample information tibble
new_exp <- mutate_samples(exp, sample = c("SI", "SII", "SIII", "SIV", "SV"))
new_exp$sample_info
#> # A tibble: 5 × 2
#>   sample group
#>   <chr>  <chr>
#> 1 SI     A    
#> 2 SII    A    
#> 3 SIII   A    
#> 4 SIV    B    
#> 5 SV     B    
new_exp$expr_mat
#>    SI SII SIII SIV SV
#> V1  1   6   11  16 21
#> V2  2   7   12  17 22
#> V3  3   8   13  18 23
#> V4  4   9   14  19 24
#> V5  5  10   15  20 25

# Modify the `variable` column in variable information tibble
new_exp <- mutate_variables(exp, variable = c("VI", "VII", "VIII", "VIV", "VV"))
new_exp$var_info
#> # A tibble: 5 × 2
#>   variable type 
#>   <chr>    <chr>
#> 1 VI       X    
#> 2 VII      X    
#> 3 VIII     Y    
#> 4 VIV      Y    
#> 5 VV       Y    
new_exp$expr_mat
#>      S1 S2 S3 S4 S5
#> VI    1  6 11 16 21
#> VII   2  7 12 17 22
#> VIII  3  8 13 18 23
#> VIV   4  9 14 19 24
#> VV    5 10 15 20 25