Select columns of the variable information tibble

Select columns of the variable information tibble of an Experiment object using <tidy-select> syntax. For example, select_variables(exp, type) will return a new Experiment object with a var_info tibble having only the variable and type columns. (Note that the variable column is always included automatically.)

Usage

select_variables(exp, ...)

Arguments

exp

The Experiment object.

...

<tidy-select> syntax for selecting columns.

Value

A new Experiment object.

Details

Instead of Experiment$select_variables(), which updates the Experiment object in place, this function returns a new Experiment object with var_info tibble updated. The original Experiment object will not be altered. Besides, as the first argument is the Experiment object, this function could be used in a magrittr pipe.

You should not, and don't need to, select or de-select the variable column. The variable column is always included automatically.